requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
